chore: M3: SecurityConfig 워크북 기준 정리 및 OAuth 소셜 UID 타입 캐스팅 수정
- SecurityConfig: 디버깅 과정에서 변경된 설정을 워크북 기준으로 복원
- CustomOAuthService: 카카오 id 속성 Long 강제 캐스팅 제거 (타입 불일치 방지)

+++ week09/src/main/java/com/example/week09/global/security/service/CustomOAuthService.java
@@ -41,7 +41,7 @@ public class CustomOAuthService extends DefaultOAuth2UserService {
             providerId = SocialType.valueOf(
                     userRequest.getClientRegistration().getRegistrationId().toUpperCase()
             );
-            socialUid = String.valueOf((Long) oAuthMember.getAttribute("id"));
+            socialUid = String.valueOf(oAuthMember.getAttribute("id"));
         } catch (IllegalArgumentException e) {
             throw new MemberException(MemberErrorCode.NOT_SUPPORT_SOCIAL_PROVIDER);
         }
